Content_Creation/requirements.txt
streamlit==1.28.1
fastmcp==0.1.0
openai==1.3.7
requests==2.31.0
python-dotenv==1.0.0
markdown==3.5.1
beautifulsoup4==4.12.2
lxml==4.9.3 